Move all price logic to the utils (#882)
* Move all price logic to an utils

* Improve name, docs, and move the cancelable function
