+++ src/components/Toggle/index.tsx
@@ -1,4 +1,4 @@
-import { WithClassName } from '@src/custom/types'
+import { WithClassName } from 'types'
 import { Trans } from '@lingui/macro'
 import React from 'react'
 import styled from 'styled-components/macro'
