+++ src/custom/utils/operator/index.ts
@@ -1,9 +1,8 @@
 import { SupportedChainId as ChainId } from 'constants/chains'
 import { getSigningSchemeApiValue, OrderCreation, OrderCancellation } from 'utils/signatures'
-import { APP_ID, PRICE_API_TIMEOUT_MS } from 'constants/index'
-import { registerOnWindow, withTimeout } from '../misc'
+import { APP_ID } from 'constants/index'
+import { registerOnWindow } from '../misc'
 import { isDev } from '../environments'
-import { FeeInformation, PriceInformation } from 'state/price/reducer'
 import OperatorError, { ApiErrorCodeDetails, ApiErrorCodes, ApiErrorObject } from 'utils/operator/errors/OperatorError'
 import QuoteError, {
   GpQuoteErrorCodes,
@@ -12,7 +11,7 @@ import QuoteError, {
   GpQuoteErrorDetails,
 } from 'utils/operator/errors/QuoteError'
 import { toErc20Address } from 'utils/tokens'
-import { getPriceQuote as getPriceQuoteParaswap } from 'utils/paraswap'
+import { FeeInformation, FeeQuoteParams, PriceInformation, PriceQuoteParams } from '../price'
 
 function getOperatorUrl(): Partial<Record<ChainId, string>> {
   if (isDev) {
@@ -165,20 +164,6 @@ export async function sendSignedOrderCancellation(params: OrderCancellationParam
   console.log('[utils:operator] Cancelled order', cancellation.orderUid, chainId)
 }
 
-export type FeeQuoteParams = Pick<OrderMetaData, 'sellToken' | 'buyToken' | 'kind'> & {
-  amount: string
-  chainId: ChainId
-  fromDecimals: number
-  toDecimals: number
-}
-
-export type PriceQuoteParams = Omit<FeeQuoteParams, 'sellToken' | 'buyToken'> & {
-  baseToken: string
-  quoteToken: string
-  fromDecimals: number
-  toDecimals: number
-}
-
 const UNHANDLED_QUOTE_ERROR: GpQuoteErrorObject = {
   errorType: GpQuoteErrorCodes.UNHANDLED_ERROR,
   description: GpQuoteErrorDetails.UNHANDLED_ERROR,
@@ -251,19 +236,5 @@ export async function getOrder(chainId: ChainId, orderId: string): Promise<Order
     throw new OperatorError(UNHANDLED_ORDER_ERROR)
   }
 }
-
-export async function getBestPriceQuote(params: PriceQuoteParams) {
-  // Get price from all API: Gpv2 and Paraswap
-  const pricePromise = withTimeout(getPriceQuote(params), PRICE_API_TIMEOUT_MS, 'GPv2: Get Price API')
-  const paraSwapPricePromise = withTimeout(
-    getPriceQuoteParaswap(params),
-    PRICE_API_TIMEOUT_MS,
-    'Paraswap: Get Price API'
-  )
-
-  // Get results from API queries
-  return Promise.allSettled([pricePromise, paraSwapPricePromise])
-}
-
 // Register some globals for convenience
 registerOnWindow({ operator: { getFeeQuote, getOrder, sendSignedOrder, apiGet: _get, apiPost: _post } })
